Updated to make user-management a NOP on Windows
Also added validation for arguments
